requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==3.2.20
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.8
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
factory